Create high level useData hook
