+++ src/hooks/data/useDataSearch.jsx
@@ -1,7 +1,7 @@
 import { useState, useCallback, useEffect } from "react";
-import { filterItemsByName } from "../utils/filterData";
+import { filterItemsByName } from "../../utils/filterData";
 
-export const useSearch = (data) => {
+export const useDataSearch = (data) => {
   const [searchData, setSearchData] = useState(data);
 
   const searchByName = useCallback(
